before keycloak fix merge

+++ src/main/java/sample/project/Controller/JobPostController.java
@@ -3,7 +3,6 @@ package sample.project.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.server.ResponseStatusException;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -12,11 +11,14 @@ import sample.project.DTO.request.JobPostRequest;
 import sample.project.DTO.request.UpdateJobPost;
 import sample.project.DTO.response.JobApplicationResponse;
 import sample.project.DTO.response.JobpostResponse;
+import sample.project.DTO.response.ServiceResponse;
 import sample.project.Model.Company;
 import sample.project.Model.User;
 import sample.project.Service.JobPostService;
+import sample.project.Service.UserService;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -35,50 +37,65 @@ import org.springframework.web.bind.annotation.GetMapping;
 public class JobPostController {
 
     private final JobPostService jobPostService;
+    private final UserService userService;
 
     @PostMapping("/create")
     @PreAuthorize("hasRole('COMPANY')")
-    public ResponseEntity<JobpostResponse> createJobPost(@Valid @RequestBody JobPostRequest req,
-            @AuthenticationPrincipal User currentUser) {
+    public ResponseEntity<?> createJobPost(@Valid @RequestBody JobPostRequest req,
+            @AuthenticationPrincipal User user) {
 
-        Company company = currentUser.getCompany();
+        Company company = user.getCompany();
         if (!company.getName().equals(req.getCompanyName())) {
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN,
-                    "Access denied , you can only create post for your company");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Access denied , you can only create post for your company");
         }
-        JobpostResponse jobPost = jobPostService.postJob(req);
-
-        return ResponseEntity.ok().body(jobPost);
+        ServiceResponse<JobpostResponse> jobPost = jobPostService.postJob(req);
+        if (!jobPost.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(jobPost.getMessage());
+        }
+        return ResponseEntity.ok().body(jobPost.getData());
     }
 
     @PostMapping("/update/{jobpostID}")
     @PreAuthorize("hasRole('COMPANY')")
-    public ResponseEntity<JobpostResponse> updateJobPost(@RequestBody UpdateJobPost req, @PathVariable Long jobpostID,
-            @AuthenticationPrincipal User currentUser) {
 
-        JobpostResponse existingjobPost = jobPostService.getJobPostById(jobpostID);
-        Company company = currentUser.getCompany();
+    public ResponseEntity<?> updateJobPost(@RequestBody UpdateJobPost req, @PathVariable Long jobpostID,
+            @AuthenticationPrincipal User user) {
 
-        boolean isSameCompany = company.getName().equals(existingjobPost.getCompanyName());
+        ServiceResponse<JobpostResponse> existingjobPost = jobPostService.getJobPostById(jobpostID);
+        if (!existingjobPost.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(existingjobPost.getMessage());
+        }
+        Company company = user.getCompany();
+
+        boolean isSameCompany = company.getName().equals(existingjobPost.getData().getCompanyName());
         if (!isSameCompany) {
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Access denied");
         }
 
-        JobpostResponse jobPost = jobPostService.updateJobPost(req, jobpostID);
-
-        return ResponseEntity.ok().body(jobPost);
+        ServiceResponse<JobpostResponse> jobPost = jobPostService.updateJobPost(req, jobpostID);
+        if (!jobPost.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(jobPost.getMessage());
+        }
+        return ResponseEntity.ok().body(jobPost.getData());
     }
 
     @DeleteMapping("/{jobpostID}")
     @PreAuthorize("hasRole('COMPANY')")
     public ResponseEntity<String> deleteJobPost(@PathVariable Long jobpostID,
-            @AuthenticationPrincipal User currentUser) {
-        JobpostResponse existingjobPost = jobPostService.getJobPostById(jobpostID);
-        Company company = currentUser.getCompany();
+            @AuthenticationPrincipal User user) {
 
-        boolean isSameCompany = company.getName().equals(existingjobPost.getCompanyName());
+        ServiceResponse<JobpostResponse> existingjobPost = jobPostService.getJobPostById(jobpostID);
+        if (!existingjobPost.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(existingjobPost.getMessage());
+        }
+        Company company = user.getCompany();
+
+        boolean isSameCompany = company.getName().equals(existingjobPost.getData().getCompanyName());
         if (!isSameCompany) {
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Access denied");
         }
         jobPostService.deleteJobPost(jobpostID);
 
@@ -87,77 +104,112 @@ public class JobPostController {
 
     @GetMapping("/{jobpostID}/jobApplication")
     @PreAuthorize("hasRole('COMPANY')")
-    public ResponseEntity<List<JobApplicationResponse>> getAllJobApplications(@PathVariable Long jobpostID,
-            @AuthenticationPrincipal User currentUser) {
-        JobpostResponse existingjobPost = jobPostService.getJobPostById(jobpostID);
-        Company company = currentUser.getCompany();
 
-        boolean isSameCompany = company.getName().equals(existingjobPost.getCompanyName());
-        if (!isSameCompany) {
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+    public ResponseEntity<?> getAllJobApplications(@PathVariable Long jobpostID,
+            @AuthenticationPrincipal User user) {
+
+        ServiceResponse<JobpostResponse> existingjobPost = jobPostService.getJobPostById(jobpostID);
+        if (!existingjobPost.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(existingjobPost.getMessage());
         }
-        List<JobApplicationResponse> jobApplications = jobPostService.getJobApplications(jobpostID);
+        Company company = user.getCompany();
 
-        return ResponseEntity.ok().body(jobApplications);
+        boolean isSameCompany = company.getName().equals(existingjobPost.getData().getCompanyName());
+        if (!isSameCompany) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Access denied");
+        }
+        ServiceResponse<List<JobApplicationResponse>> jobApplications = jobPostService.getJobApplications(jobpostID);
+        if (!jobApplications.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(jobApplications.getMessage());
+        }
+
+        return ResponseEntity.ok().body(jobApplications.getData());
 
     }
 
     @GetMapping("/{jobPostID}/jobApplication/{jobApplicationID}")
     @PreAuthorize("hasRole('COMPANY')")
-    public ResponseEntity<JobApplicationResponse> getJobApplication(@PathVariable Long jobPostID,
+    public ResponseEntity<?> getJobApplication(@PathVariable Long jobPostID,
             @PathVariable Long jobApplicationID,
-            @AuthenticationPrincipal User currentUser) {
-        JobpostResponse existingjobPost = jobPostService.getJobPostById(jobPostID);
-        Company company = currentUser.getCompany();
+            @AuthenticationPrincipal User user) {
 
-        boolean isSameCompany = company.getName().equals(existingjobPost.getCompanyName());
-        if (!isSameCompany) {
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+        ServiceResponse<JobpostResponse> existingjobPost = jobPostService.getJobPostById(jobPostID);
+        if (!existingjobPost.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(existingjobPost.getMessage());
         }
-        JobApplicationResponse jobApplications = jobPostService.getJobApplication(jobPostID, jobApplicationID);
+        Company company = user.getCompany();
 
-        return ResponseEntity.ok().body(jobApplications);
+        boolean isSameCompany = company.getName().equals(existingjobPost.getData().getCompanyName());
+        if (!isSameCompany) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Access denied");
+        }
+        ServiceResponse<JobApplicationResponse> jobApplications = jobPostService.getJobApplication(jobPostID,
+                jobApplicationID);
+
+        if (!jobApplications.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(jobApplications.getMessage());
+        }
+
+        return ResponseEntity.ok().body(jobApplications.getData());
 
     }
 
     @PutMapping("/{jobPostID}/jobApplication/{jobApplicationID}")
     @PreAuthorize("hasRole('COMPANY')")
-    public ResponseEntity<JobApplicationResponse> updateJobApplicationStatus(@PathVariable Long jobPostID,
+    public ResponseEntity<?> updateJobApplicationStatus(@PathVariable Long jobPostID,
             @PathVariable Long jobApplicationID,
             @RequestBody JobApplicationUpdateRequest req,
-            @AuthenticationPrincipal User currentUser) {
-        JobpostResponse existingjobPost = jobPostService.getJobPostById(jobPostID);
-        Company company = currentUser.getCompany();
+            @AuthenticationPrincipal User user) {
 
-        boolean isSameCompany = company.getName().equals(existingjobPost.getCompanyName());
-        if (!isSameCompany) {
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+        ServiceResponse<JobpostResponse> existingjobPost = jobPostService.getJobPostById(jobPostID);
+        if (!existingjobPost.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(existingjobPost.getMessage());
         }
-        JobApplicationResponse jobApplications = jobPostService.updateJobApplication(jobPostID, jobApplicationID, req);
+        Company company = user.getCompany();
 
-        return ResponseEntity.ok().body(jobApplications);
+        boolean isSameCompany = company.getName().equals(existingjobPost.getData().getCompanyName());
+        if (!isSameCompany) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Access denied");
+        }
+        ServiceResponse<JobApplicationResponse> jobApplications = jobPostService.updateJobApplication(jobPostID,
+                jobApplicationID, req);
+        if (!jobApplications.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(jobApplications.getMessage());
+        }
+        return ResponseEntity.ok().body(jobApplications.getData());
 
     }
 
     @GetMapping("/company")
-    public ResponseEntity<List<JobpostResponse>> getAllJobPostsCompany(@AuthenticationPrincipal User currentUser,
+
+    public ResponseEntity<?> getAllJobPostsCompany(@AuthenticationPrincipal User user,
             @RequestParam(required = false) Integer salaryMin,
             @RequestParam(required = false) Integer salaryMax,
             @RequestParam(required = false) String date,
             @RequestParam(required = false) String search,
             @RequestParam(required = false) String sort) {
 
-        List<JobpostResponse> jobPosts = jobPostService.getAllJobPostsCompany(currentUser.getCompany().getName(),
+        ServiceResponse<List<JobpostResponse>> jobPosts = jobPostService.getAllJobPostsCompany(
+                user.getCompany().getName(),
                 salaryMin, salaryMax, date, sort, search);
-        return ResponseEntity.ok().body(jobPosts);
+
+        if (!jobPosts.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(jobPosts.getMessage());
+        }
+        return ResponseEntity.ok().body(jobPosts.getData());
 
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<JobpostResponse> getJobPostById(@PathVariable Long id) {
-        JobpostResponse jobPost = jobPostService.getJobPostById(id);
-
-        return ResponseEntity.ok().body(jobPost);
+    public ResponseEntity<?> getJobPostById(@PathVariable Long id) {
+        ServiceResponse<JobpostResponse> jobPost = jobPostService.getJobPostById(id);
+        if (!jobPost.isSuccess()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(jobPost.getMessage());
+        }
+        return ResponseEntity.ok().body(jobPost.getData());
 
     }
